requirements.txt
altgraph==0.17.2
auto-py-to-exe==2.18.0
beautifulsoup4==4.10.0
bottle==0.12.19
bottle-websocket==0.2.9
Brotli==1.0.9
bs4==0.0.1
cffi==1.15.0
colorama==0.4.4
cssselect2==0.5.0
cycler==0.11.0
Eel==0.12.4
fonttools==4.29.1
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
html2image==2.0.1
html5lib==1.1
imgkit==1.2.2
imutils==0.5.4
kiwisolver==1.4.2
matplotlib==3.5.1
numpy==1.22.2
opencv-contrib-python==4.5.5.62
packaging==21.3
pefile==2021.9.3
Pillow==9.0.1
pycparser==2.21
pydyf==0.1.2
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.2
pyparsing==3.0.8
pyphen==0.12.0
python-dateutil==2.8.2
pywin32-ctypes==0.2.0
six==1.16.0
soupsieve==2.3.1
texttohtml==1.0
tinycss2==1.1.1
tqdm==4.64.0
weasyprint==54.2
webencodings==0.5.1
whichcraft==0.6.1
wkhtmltopdf==0.2
zope.event==4.5.0
zope.interface==5.4.0
zopfli==0.2.1
